Working RTM and implementation of laplacian filter

+++ modelling_utils.h
@@ -5,4 +5,4 @@ struct b{
 	int nb;
 };
 void velextension(float* vex,float* v,int nx,int nz,struct b border,int nxx,int nzz);
-float fricker(float t,float freq);
+double fricker(float t,float freq);
